rearrange code to use Interrupt.New split machine_esp32c3.go into module related files.
